Add files via upload
The main objective of this application is to raise awareness and inform the public in general to preserve the intangible cultural heritage from the extreme weather conditions that affect the whole world (text, video, images, quiz, maps). The content was collected over the web with examples of worldwide scope for this application. In this way it is expected that the social interpretation, the social role and the spatial dimension, of the morals, customs, knowledge and information, the dances, the oral traditions and those that exist to this day will be saved. It is believed that through this application we will be able to raise the awareness of a large part of the population and arouse the interest of people who do not know what part of culture is called intangible cultural heritage. As a result, the work aims to create a “new” tool, as it is “built” from the “start” and the experiences of citizens are collected over time.